Updates on the login (html & php) to perform user authentication
The list is long… I altered the length of the password field on the
user table to mach the length of the sha256 hash (now it’s 64 chars
long). The php code is performing auth and redirects. Once again, the
login information for our current user is user/Test123!

I organized the sql scripts in a directory and created two files for
ddl and dml statements so we don’t need to drop/rebuild the tables
every time something changes. Also, I moved the login tables from that
separate script to our main one.